+++ client/tws_foundation_client/lib/src/services/business/bases/locations_service_base.dart
@@ -1,20 +1,19 @@
 import 'package:tws_foundation_client/src/services/tws_service_base.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
-abstract class SectionsServiceBase extends TWSServiceBase {
-  SectionsServiceBase(
+abstract class LocationsServiceBase extends TWSServiceBase {
+  LocationsServiceBase(
     super.host,
     super.servicePath, {
     super.client,
   });
 
   /// Transaction to generate a set view object.
-  Effect<SetViewOut<Section>> view(SetViewOptions<Section> options, String auth);
+  Effect<SetViewOut<Location>> view(SetViewOptions<Location> options, String auth);
 
   /// Transaction to create a set view object.
-  Effect<SetBatchOut<Section>> create(List<Section> sections, String auth);
+  Effect<SetBatchOut<Location>> create(List<Location> locations, String auth);
 
   /// Transaction to update a set object.
-  Effect<RecordUpdateOut<Section>> update(Section section, String auth);
-
+  Effect<RecordUpdateOut<Location>> update(Location location, String auth);
 }