+++ client/tws_foundation_client/test/definition/locations_service_test.dart
@@ -5,28 +5,28 @@ import 'package:test/test.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
 void main() {
-  late SectionsServiceBase service;
-  late SetViewOut<Section> viewMock;
-  late SetBatchOut<Section> createMock;
-  late RecordUpdateOut<Section> updateMock;
-  late SetViewOptions<Section> options;
-  late List<Section> sections;
+  late LocationsServiceBase service;
+  late SetViewOut<Location> viewMock;
+  late SetBatchOut<Location> createMock;
+  late RecordUpdateOut<Location> updateMock;
+  late SetViewOptions<Location> options;
+  late List<Location> locations;
   setUp(
     () {
       List<SetViewOrderOptions> noOrderigns = <SetViewOrderOptions>[];
-      options = SetViewOptions<Section>(false, 10, 1, null, noOrderigns, <SetViewFilterNodeInterface<Section>>[]);
-      viewMock = SetViewOut<Section>(<Section>[], 1, DateTime.now(), 3, 0, 20);
-      createMock = SetBatchOut<Section>(<Section>[], <SetOperationFailure<Section>>[], 0, 0, 0, false);
-      updateMock = RecordUpdateOut<Section>(Section.a(), Section.a());
-      sections = <Section>[
-        Section.a(),
+      options = SetViewOptions<Location>(false, 10, 1, null, noOrderigns, <SetViewFilterNodeInterface<Location>>[]);
+      viewMock = SetViewOut<Location>(<Location>[], 1, DateTime.now(), 3, 0, 20);
+      createMock = SetBatchOut<Location>(<Location>[], <SetOperationFailure<Location>>[], 0, 0, 0, false);
+      updateMock = RecordUpdateOut<Location>(Location.a(), Location.a());
+      locations = <Location>[
+        Location.a(),
       ];
       Client mockClient = MockClient(
         (Request request) async {
           JObject jObject = switch (request.url.pathSegments.last) {
-            'view' => SuccessFrame<SetViewOut<Section>>('qTracer', viewMock).encode(),
-            'create' => SuccessFrame<SetBatchOut<Section>>('qTracer', createMock).encode(),
-            'update' => SuccessFrame<RecordUpdateOut<Section>>('qTracer', updateMock).encode(),
+            'view' => SuccessFrame<SetViewOut<Location>>('qTracer', viewMock).encode(),
+            'create' => SuccessFrame<SetBatchOut<Location>>('qTracer', createMock).encode(),
+            'update' => SuccessFrame<RecordUpdateOut<Location>>('qTracer', updateMock).encode(),
             _ => <String, dynamic>{},
           };
 
@@ -37,18 +37,18 @@ void main() {
       service = TWSFoundationSource(
         true,
         client: mockClient,
-      ).sections;
+      ).locations;
     },
   );
 
   test(
     'View',
     () async {
-      MainResolver<SetViewOut<Section>> fact = await service.view(options, '');
+      MainResolver<SetViewOut<Location>> fact = await service.view(options, '');
 
       bool passed = false;
       fact.resolve(
-        decoder: (JObject json) => SetViewOut<Section>.des(json, Section.des),
+        decoder: (JObject json) => SetViewOut<Location>.des(json, Location.des),
         onConnectionFailure: () {},
         onFailure: (FailureFrame failure, int status) {
           assert(false, 'server returned a success $status');
@@ -56,10 +56,10 @@ void main() {
         onException: (Object exception, StackTrace trace) {
           assert(false, 'server returned a success');
         },
-        onSuccess: (SuccessFrame<SetViewOut<Section>> success) {
+        onSuccess: (SuccessFrame<SetViewOut<Location>> success) {
           passed = true;
 
-          SetViewOut<Section> fact = success.estela;
+          SetViewOut<Location> fact = success.estela;
           expect(viewMock.page, fact.page);
           expect(viewMock.pages, fact.pages);
           expect(viewMock.records, fact.records);
@@ -74,15 +74,15 @@ void main() {
    test(
     'Create',
     () async {
-      MainResolver<SetBatchOut<Section>> fact = await service.create(sections, '');
+      MainResolver<SetBatchOut<Location>> fact = await service.create(locations, '');
       bool pased = false;
       fact.resolve(
-        decoder: (JObject json) => SetBatchOut<Section>.des(json, Section.des),
+        decoder: (JObject json) => SetBatchOut<Location>.des(json, Location.des),
         onConnectionFailure: () {},
         onFailure: (FailureFrame failure, int status) {
           throw failure;
         },
-        onSuccess: (SuccessFrame<SetBatchOut<Section>> success) {
+        onSuccess: (SuccessFrame<SetBatchOut<Location>> success) {
           pased = true;
         },
         onException: (Object exception, StackTrace trace) {
@@ -96,15 +96,15 @@ void main() {
   test(
     'Update',
     () async {
-      MainResolver<RecordUpdateOut<Section>> fact = await service.update(Section.a(), '');
+      MainResolver<RecordUpdateOut<Location>> fact = await service.update(Location.a(), '');
       bool pased = false;
       fact.resolve(
-        decoder: (JObject json) => RecordUpdateOut<Section>.des(json, Section.des),
+        decoder: (JObject json) => RecordUpdateOut<Location>.des(json, Location.des),
         onConnectionFailure: () {},
         onFailure: (FailureFrame failure, int status) {
           throw failure;
         },
-        onSuccess: (SuccessFrame<RecordUpdateOut<Section>> success) {
+        onSuccess: (SuccessFrame<RecordUpdateOut<Location>> success) {
           pased = true;
         },
         onException: (Object exception, StackTrace trace) {