+++ schemas/CURRENT/4. addess_nulleable_in_locations.sql
@@ -1,0 +1,8 @@
+USE [TWS Business];
+
+-- Changing [Address] in Locations to nulleable.
+IF EXISTS (SELECT * FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_NAME = 'Locations' AND COLUMN_NAME = 'Address')
+BEGIN
+    ALTER TABLE Locations
+	ALTER COLUMN [Address] INT;
+END