+++ server/TWS Foundation/Server.Quality/Suit/Controllers/Business/Q_LocationsController.cs
@@ -1,11 +1,12 @@
-﻿using System.Net;
-
+﻿
 using CSM_Foundation.Core.Utils;
 using CSM_Foundation.Database.Models.Options;
 using CSM_Foundation.Database.Models.Out;
 using CSM_Foundation.Server.Records;
-
 using Microsoft.AspNetCore.Mvc.Testing;
+
+using System;
+using System.Net;
 
 using TWS_Business.Sets;
 
@@ -16,46 +17,14 @@ using TWS_Foundation.Middlewares.Frames;
 using TWS_Foundation.Quality.Bases;
 
 using Account = TWS_Foundation.Quality.Secrets.Account;
-using View = CSM_Foundation.Database.Models.Out.SetViewOut<TWS_Business.Sets.Section>;
-
+using View = CSM_Foundation.Database.Models.Out.SetViewOut<TWS_Business.Sets.Location>;
 
 namespace TWS_Foundation.Quality.Suit.Controllers.Business;
-public class Q_SectionsController
-    : BQ_CustomServerController<Section> {
-
-    private class Frame : SuccessFrame<View> { }
-
-
-    public Q_SectionsController(WebApplicationFactory<Program> hostFactory)
-        : base("Sections", hostFactory) {
+public class Q_LocationsController
+    : BQ_CustomServerController<Location> {
+    public Q_LocationsController(WebApplicationFactory<Program> hostFactory)
+        : base("Locations", hostFactory) {
     }
-
-    protected override Section MockFactory(string RandomSeed) {
-        int random = new Random().Next();
-        decimal cords = (random % 100) + 0.1m;
-        return new() {
-            Status = 1,
-            Name = "name" + RandomSeed,
-            Yard = 0,
-            Capacity = new Random().Next(100),
-            Ocupancy = new Random().Next(100),
-            LocationNavigation = new() {
-                Status = 1,
-                Name = "name " + RandomSeed,
-                AddressNavigation = new() {
-                    Country = "USA",
-                    City = "city " + RandomSeed,
-                },
-                WaypointNavigation = new() {
-                    Status = 1,
-                    Longitude = cords,
-                    Latitude = cords,
-                    Altitude = cords,
-                }
-            },
-        };
-    }
-
     protected override async Task<string> Authentication() {
         (HttpStatusCode Status, SuccessFrame<Session> Response) = await XPost<SuccessFrame<Session>, Credentials>("Security/Authenticate", new Credentials {
             Identity = Account.Identity,
@@ -64,6 +33,25 @@ public class Q_SectionsController
         });
 
         return Status != HttpStatusCode.OK ? throw new ArgumentNullException(nameof(Status)) : Response.Estela.Token.ToString();
+    }
+
+    protected override Location MockFactory(string RandomSeed) {
+        int random = new Random().Next();
+        decimal cords = (random % 100) + 0.1m;
+        return new() {
+            Status = 1,
+            Name = "name " + RandomSeed,
+            AddressNavigation = new() {
+                Country = "USA",
+                City = "city " + RandomSeed,
+            },
+            WaypointNavigation = new() {
+                Status = 1,
+                Longitude = cords,
+                Latitude = cords,
+                Altitude = cords,
+            }
+        };
     }
 
     [Fact]
@@ -84,7 +72,7 @@ public class Q_SectionsController
 
     [Fact]
     public async Task Create() {
-        List<Section> mockList = [];
+        List<Location> mockList = [];
         string testTag = Guid.NewGuid().ToString()[..2];
 
         for (int i = 0; i < 3; i++) {
@@ -93,7 +81,7 @@ public class Q_SectionsController
         }
 
         (HttpStatusCode Status, GenericFrame response) = await Post("Create", mockList, true);
-        SetBatchOut<Section> estela = Framing<SuccessFrame<SetBatchOut<Section>>>(response).Estela;
+        SetBatchOut<Location> estela = Framing<SuccessFrame<SetBatchOut<Location>>>(response).Estela;
         Assert.Equal(HttpStatusCode.OK, Status);
         Assert.Empty(estela.Failures);
     }
@@ -103,15 +91,15 @@ public class Q_SectionsController
         #region First (Correctly creates when doesn't exist)
         {
             string testTag = Guid.NewGuid().ToString()[..3];
-            Section mock = MockFactory(testTag);
+            Location mock = MockFactory(testTag);
             (HttpStatusCode Status, GenericFrame Respone) = await Post("Update", mock, true);
 
             Assert.Equal(HttpStatusCode.OK, Status);
-            RecordUpdateOut<Section> creationResult = Framing<SuccessFrame<RecordUpdateOut<Section>>>(Respone).Estela;
+            RecordUpdateOut<Location> creationResult = Framing<SuccessFrame<RecordUpdateOut<Location>>>(Respone).Estela;
 
             Assert.Null(creationResult.Previous);
 
-            Section updated = creationResult.Updated;
+            Location updated = creationResult.Updated;
             Assert.True(updated.Id > 0);
         }
         #endregion
@@ -120,20 +108,20 @@ public class Q_SectionsController
         {
             #region generate a new record
             string testTag = Guid.NewGuid().ToString()[..3];
-            Section mock = MockFactory(testTag);
+            Location mock = MockFactory(testTag);
 
             (HttpStatusCode Status, GenericFrame Response) = await Post("Update", mock, true);
 
             Assert.Equal(HttpStatusCode.OK, Status);
 
-            RecordUpdateOut<Section> creationResult = Framing<SuccessFrame<RecordUpdateOut<Section>>>(Response).Estela;
+            RecordUpdateOut<Location> creationResult = Framing<SuccessFrame<RecordUpdateOut<Location>>>(Response).Estela;
             Assert.Null(creationResult.Previous);
 
-            Section creationRecord = creationResult.Updated;
+            Location creationRecord = creationResult.Updated;
             Assert.Multiple([
                 () => Assert.True(creationRecord.Id > 0),
-                () => Assert.Equal(mock.Name, creationRecord.Name),
-                () => Assert.Equal(mock.LocationNavigation!.Name, creationRecord.LocationNavigation!.Name),
+                () => Assert.Equal(mock.AddressNavigation!.City, creationRecord.AddressNavigation!.City),
+                () => Assert.Equal(mock.WaypointNavigation!.Longitude, creationRecord.WaypointNavigation!.Longitude),
             ]);
             #endregion
 
@@ -142,27 +130,28 @@ public class Q_SectionsController
             string updatedTag = "UPDTE";
             mock = creationRecord;
 
-            mock.Name = updatedTag + RandomUtils.String(7);
-            mock.LocationNavigation!.Name = updatedTag + RandomUtils.String(7);
+            mock.AddressNavigation!.City = updatedTag + RandomUtils.String(7);
+            mock.WaypointNavigation!.Latitude = 99.99m;
             (HttpStatusCode Status, GenericFrame Response) updateResponse = await Post("Update", mock, true);
 
             Assert.Equal(HttpStatusCode.OK, updateResponse.Status);
-            RecordUpdateOut<Section> updateResult = Framing<SuccessFrame<RecordUpdateOut<Section>>>(updateResponse.Response).Estela;
+            RecordUpdateOut<Location> updateResult = Framing<SuccessFrame<RecordUpdateOut<Location>>>(updateResponse.Response).Estela;
 
             Assert.NotNull(updateResult.Previous);
 
-            Section updateRecord = updateResult.Updated;
-            Section previousRecord = updateResult.Previous;
+            Location updateRecord = updateResult.Updated;
+            Location previousRecord = updateResult.Previous;
             Assert.Multiple([
                 () => Assert.Equal(creationRecord.Id, updateRecord.Id),
-                () => Assert.Equal(creationRecord.Status, updateRecord.Status),
-                () => Assert.NotEqual(previousRecord.Name, updateRecord.Name),
-                () => Assert.NotEqual(previousRecord.LocationNavigation!.Name, updateRecord.LocationNavigation!.Name),
+                () => Assert.Equal(creationRecord.AddressNavigation!.Country, updateRecord.AddressNavigation!.Country),
+                () => Assert.NotEqual(previousRecord.AddressNavigation!.City, updateRecord.AddressNavigation!.City),
+                () => Assert.NotEqual(previousRecord.WaypointNavigation!.Latitude, updateRecord.WaypointNavigation!.Latitude),
             ]);
             #endregion
         }
         #endregion
     }
-
-
 }
+
+
+