+++ server/TWS Foundation/TWS Foundation/Controllers/Business/SectionsController.cs
@@ -10,15 +10,25 @@ using TWS_Foundation.Authentication;
 
 namespace TWS_Foundation.Controllers.Business;
 
-[ApiController, Route("[Controller]/[Action]")]
+[ApiController, Route("[Controller]")]
 public class SectionsController : ControllerBase {
     private readonly ISectionsService Service;
     public SectionsController(ISectionsService service) {
         Service = service;
     }
 
-    [HttpPost(), Auth("", "")]
+    [HttpPost("[Action]"), Auth("", "")]
     public async Task<IActionResult> View(SetViewOptions<Section> Options) {
         return Ok(await Service.View(Options));
+    }
+
+    [HttpPost("[Action]"), Auth("", "")]
+    public async Task<IActionResult> Create(Section[] sections) {
+        return Ok(await Service.Create(sections));
+    }
+
+    [HttpPost("[Action]"), Auth("", "")]
+    public async Task<IActionResult> Update(Section section) {
+        return Ok(await Service.Update(section));
     }
 }